backend/requirements.txt
# Основные зависимости
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# База данных
psycopg2-binary==2.9.9
supabase==1.1.1
sqlalchemy==2.0.23
alembic==1.13.1  # опционально

# Работа с данными
pandas==2.1.4
openpyxl==3.1.2
numpy==1.26.2

# Кэширование и очереди
redis==5.0.1
pika==1.3.2
celery==5.3.4

# ИИ и NLP
openai==1.12.0
langchain==0.1.0  # опционально
tiktoken==0.5.2

# Валидация
pydantic==2.5.0
pydantic-settings==2.1.0

# Утилиты
python-json-logger==2.0.7
structlog==23.2.0
requests==2.31.0
aiohttp==3.9.1

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
requests-mock==1.11.0

# Dev dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.0
pre-commit==3.5.0